chore: refactoring code
A new spotify helper file has been added. The relevant functions have
been removed from the route-based file and placed there for better
readability.
